refactor: trocear el form de viviendas en secciones + adoptar FormField (F8/X5)
[Plan 06 · F8/X5]
- F8: PropertyCreatePage deja de ser un monolito; las secciones de campos se
  extraen a components/Admin/Properties/sections/: BasicInfoSection,
  FeaturesSection, LocationSection, ClassificationSection. El padre queda como
  orquestador (estado, submit, validacion, autoguardado, popups, imagenes).
  getPlainTextFromHtml se movio a utils/htmlText.js.
- X5: cada campo de las secciones usa el componente FormField (label + control
  + ayuda + un unico estilo de error), unificando los .form-group divergentes.
Build verde; sin referencias colgantes.
